Added new aliases and wezterm updates
* Aliases for fzf process killing and asdf install and clean
* Wezterm update to fix cursor rendering in neovim

+++ .aliases.sh
@@ -7,6 +7,11 @@ alias weather="curl 'wttr.in/copenhagen?m'"
 alias ftldr="tldr --list | fzf --preview 'tldr {} --color always' | xargs tldr"
 alias fta="tmux_switcher"
 alias ftk="tmux_kill_session"
+alias kp="kill_process"
+# Mnemonic [V]ersion [M]anager [I]nstall
+alias vmi="asdf_install"
+# Mnemonic [V]ersion [M]anager [C]lean
+alias vmc="asdf_remove"
 
 if [ $(command -v eza) ]; then
     alias ls='eza'                 # just replace ls by eza and allow all other eza arguments
@@ -20,22 +25,69 @@ if [ $(command -v eza) ]; then
 fi
 
 tmux_switcher() {
-  [[ -n "$TMUX" ]] && change="switch-client" || change="attach-session"
-  if [ $1 ]; then
-    tmux $change -t "$1" 2>/dev/null || (tmux new-session -d -s $1 && tmux $change -t "$1"); return
-  fi
-  session=$(tmux list-sessions -F "#{session_name}" 2>/dev/null | fzf --exit-0) &&  tmux $change -t "$session" || echo "No sessions found."
+    [[ -n "$TMUX" ]] && change="switch-client" || change="attach-session"
+    if [ $1 ]; then
+        tmux $change -t "$1" 2>/dev/null || (tmux new-session -d -s $1 && tmux $change -t "$1"); return
+    fi
+    session=$(tmux list-sessions -F "#{session_name}" 2>/dev/null | fzf --exit-0) &&  tmux $change -t "$session" || echo "No sessions found."
 }
 
 tmux_kill_session() {
     local sessions
     sessions="$(tmux ls|fzf --exit-0 --multi)"  || return $?
     local i
-    for i in "${(f@)sessions}"
-    do
-        [[ $i =~ '([^:]*):.*' ]] && {
+    for i in "${(f@)sessions}"; do
+        if [[ $i =~ '([^:]*):.*' ]]; then
             echo "Killing $match[1]"
             tmux kill-session -t "$match[1]"
-        }
+        fi
     done
+}
+
+kill_process() {
+    local pid
+    pid=$(ps -ef | sed 1d | fzf -m | awk '{print $2}')
+
+    if [ "x$pid" != "x" ]
+    then
+        echo $pid | xargs kill -${1:-9}
+    fi
+}
+
+# Install one or more versions of specified language
+# e.g. `vmi rust` # => fzf multimode, tab to mark, enter to install
+# if no plugin is supplied (e.g. `vmi<CR>`), fzf will list them for you
+asdf_install() {
+  local lang=${1}
+
+  if [[ ! $lang ]]; then
+    lang=$(asdf plugin-list | fzf)
+  fi
+
+  if [[ $lang ]]; then
+    local versions=$(asdf list-all $lang | fzf --tac --no-sort --multi)
+    if [[ $versions ]]; then
+      for version in $(echo $versions);
+      do; asdf install $lang $version; done;
+    fi
+  fi
+}
+
+# Remove one or more versions of specified language
+# e.g. `vmi rust` # => fzf multimode, tab to mark, enter to remove
+# if no plugin is supplied (e.g. `vmi<CR>`), fzf will list them for you
+asdf_remove() {
+  local lang=${1}
+
+  if [[ ! $lang ]]; then
+    lang=$(asdf plugin-list | fzf)
+  fi
+
+  if [[ $lang ]]; then
+      local versions=$(asdf list $lang | fzf -m)
+      if [[ $versions ]]; then
+          for version in $(echo $versions);
+          do; asdf uninstall $lang $version; done;
+      fi
+  fi
 }
